Move decrypt/encrypt functions to their own modules
We will add secret sharing, so crypto module will grow in size. It makes sense to keep things in their own modules now.

It also does a bit of a refactoring to prepare to encrypt something that is not `IpaPrivateData`, but no new functionality is added by this change.
