+++ ipa-core/src/bin/crypto_util.rs
@@ -2,7 +2,7 @@ use std::fmt::Debug;
 
 use clap::{Parser, Subcommand};
 use ipa_core::{
-    cli::crypto::{decrypt_and_reconstruct, encrypt, DecryptArgs, EncryptArgs},
+    cli::crypto::{DecryptArgs, EncryptArgs},
     error::BoxError,
 };
 
@@ -24,8 +24,8 @@ enum CryptoUtilCommand {
 async fn main() -> Result<(), BoxError> {
     let args = Args::parse();
     match args.action {
-        CryptoUtilCommand::Encrypt(encrypt_args) => encrypt(&encrypt_args)?,
-        CryptoUtilCommand::Decrypt(decrypt_args) => decrypt_and_reconstruct(decrypt_args).await?,
+        CryptoUtilCommand::Encrypt(encrypt_args) => encrypt_args.encrypt()?,
+        CryptoUtilCommand::Decrypt(decrypt_args) => decrypt_args.decrypt_and_reconstruct().await?,
     }
     Ok(())
 }
